factor out session utility functions
